Add PromiseTask so we can want until a task returns some value.

+++ src/main/java/info/sunng/stages/sample/URLDownloadTask.java
@@ -24,7 +24,7 @@ public class URLDownloadTask extends AbstractRetryableTask {
     }
 
     @Override
-    protected int getNextRetryDelay() {
+    protected long getNextRetryDelay() {
         return 500;
     }
